// src/app/models/spartan-user-rule-module.ts
export class UserRuleModule {
    Id:                                  number;
    Module_Id:                           number;
    Module_Id_Spartan_Module:            any;
    Order_Shown:                         number;
    Spartan_User_Role:                   number;
    Spartan_User_Role_Spartan_User_Role: any;
    User_Rule_Module_Id:                 number;
}